Move chat related files into side_panel folder
