+++ chrome/browser/chrome_browser_interface_binders.cc
@@ -52,8 +52,8 @@
 #include "chrome/browser/ui/webui/suggest_internals/suggest_internals_ui.h"
 #include "chrome/browser/ui/webui/usb_internals/usb_internals.mojom.h"
 #include "chrome/browser/ui/webui/usb_internals/usb_internals_ui.h"
-#include "chrome/browser/ui/webui/chat/chat.mojom.h"
-#include "chrome/browser/ui/webui/chat/chat_ui.h"
+#include "chrome/browser/ui/webui/side_panel/chat/chat.mojom.h"
+#include "chrome/browser/ui/webui/side_panel/chat/chat_ui.h"
 #include "chrome/browser/web_applications/web_app_utils.h"
 #include "chrome/common/buildflags.h"
 #include "chrome/common/chrome_features.h"
