+++ chrome/browser/resources/BUILD.gn
@@ -61,6 +61,7 @@ group("resources") {
       "side_panel/history_clusters:resources",
       "side_panel/read_anything:resources",
       "side_panel/reading_list:resources",
+      "side_panel/chat:resources",
       "side_panel/shared:resources",
       "signin:resources",
       "suggest_internals:resources",
@@ -69,7 +70,6 @@ group("resources") {
       "web_app_internals:resources",
       "webui_gallery:resources",
       "whats_new:resources",
-      "chat:resources",
     ]
     if (is_chrome_branded) {
       public_deps += [
