+++ chrome/browser/resources/side_panel/chat/BUILD.gn
@@ -1,7 +1,7 @@
 import("//ui/webui/resources/tools/build_webui.gni")
 
 build_webui("build") {
-    grd_prefix = "chat"
+    grd_prefix = "side_panel_chat"
 
     static_files = [ "chat.html", "chat.css" ]
 
@@ -17,6 +17,6 @@ build_webui("build") {
         "//ui/webui/resources/mojo:build_ts",
     ]
 
-    mojo_files_deps = [ "//chrome/browser/ui/webui/chat:mojo_bindings_ts__generator" ]
-    mojo_files = [ "$root_gen_dir/chrome/browser/ui/webui/chat/chat.mojom-webui.ts",]
+    mojo_files_deps = [ "//chrome/browser/ui/webui/side_panel/chat:mojo_bindings_ts__generator" ]
+    mojo_files = [ "$root_gen_dir/chrome/browser/ui/webui/side_panel/chat/chat.mojom-webui.ts",]
 }
