+++ chrome/browser/ui/views/frame/browser_view.cc
@@ -1035,7 +1035,6 @@ BrowserView::BrowserView(std::unique_ptr<Browser> browser)
       this, is_right_aligned ? SidePanel::HorizontalAlignment::kRight
                              : SidePanel::HorizontalAlignment::kLeft));
   unified_side_panel_->SetPanelWidth(500);
-  unified_side_panel_->SetBackgroundRadii(gfx::RoundedCornersF(0.f, 0.f, 0.f, 0.f));
   left_aligned_side_panel_separator_ =
       AddChildView(std::make_unique<ContentsSeparator>());
   side_panel_rounded_corner_ =
